requirements.txt
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
drf-flex-fields==0.3.5
Markdown==3.1
pytz==2019.1
sqlparse==0.3.0
